// synapse_ui/src/generated/messages/v1/message.ts
export enum MessageTypeProto {
  MESSAGE_TYPE_PROTO_UNSPECIFIED = 0,
  MESSAGE_TYPE_PROTO_SEND_DEVICE_INFO = 1,
  MESSAGE_TYPE_PROTO_SEND_METRICS = 2,
  MESSAGE_TYPE_PROTO_ADD_PIPELINE = 3,
  MESSAGE_TYPE_PROTO_SEND_PIPELINE_TYPES = 4,
  MESSAGE_TYPE_PROTO_ADD_CAMERA = 5,
  MESSAGE_TYPE_PROTO_SET_TYPE_FOR_PIPELINE = 6,
  MESSAGE_TYPE_PROTO_SET_SETTING = 7,
  MESSAGE_TYPE_PROTO_SET_PIPELINE_INDEX = 8,
  MESSAGE_TYPE_PROTO_SET_PIPELINE_NAME = 9,
  MESSAGE_TYPE_PROTO_SET_DEFAULT_PIPELINE = 10,
  MESSAGE_TYPE_PROTO_DELETE_PIPELINE = 11,
  MESSAGE_TYPE_PROTO_LOG = 12,
  MESSAGE_TYPE_PROTO_SAVE = 13,
  MESSAGE_TYPE_PROTO_REPORT_CAMERA_PERFORMANCE = 14,
  MESSAGE_TYPE_PROTO_SET_NETWORK_SETTINGS = 15,
  MESSAGE_TYPE_PROTO_REBOOT = 16,
  MESSAGE_TYPE_PROTO_FORMAT = 17,
  MESSAGE_TYPE_PROTO_RESTART_SYNAPSE = 18,
  MESSAGE_TYPE_PROTO_RENAME_CAMERA = 19,
  MESSAGE_TYPE_PROTO_CALIBRATING = 20,
  MESSAGE_TYPE_PROTO_CALIBRATION_DATA = 21,
  MESSAGE_TYPE_PROTO_DELETE_CALIBRATION = 22,
  MESSAGE_TYPE_PROTO_SET_CAMERA_RECORDING_STATUS = 23,
  MESSAGE_TYPE_PROTO_SET_DEVICE_CONNECTION_STATUS = 24,
  MESSAGE_TYPE_PROTO_SET_PIPELINE_RESULT = 25,
  MESSAGE_TYPE_PROTO_ALERT = 26,
  UNRECOGNIZED = -1,
}
// ...
export interface MessageProto {
  /** Type of the message, indicating what kind of data it contains */
  type: MessageTypeProto;
  /** Information about the device */
  deviceInfo?: DeviceInfoProto | undefined;
  /** Metrics related to hardware performance */
  hardwareMetrics?: HardwareMetricsProto | undefined;
  /** Information about the camera */
  cameraInfo?: CameraProto | undefined;
  /** Information about the pipeline */
  pipelineInfo?: PipelineProto | undefined;
  /** Message to set the type for a pipeline */
  setPipelineType?: SetPipelineTypeMessageProto | undefined;
  /** Message to set a setting for a pipeline */
  setPipelineSetting?: SetPipleineSettingMessageProto | undefined;
  /** Message to set the index of a pipeline */
  setPipelineIndex?: SetPipelineIndexMessageProto | undefined;
  /** Message to set the name of a pipeline */
  setPipelineName?: SetPipelineNameMessageProto | undefined;
  /** Message to set the default pipeline */
  setDefaultPipeline?: SetDefaultPipelineMessageProto | undefined;
  removePipeline?: RemovePipelineMessageProto | undefined;
  /** Log message containing log data */
  log?: LogMessageProto | undefined;
  /** Performance data for a camera */
  cameraPerformance?: CameraPerformanceProto | undefined;
  /** Message to set network settings */
  setNetworkSettings?: SetNetworkSettingsProto | undefined;
  /** Message to rename a camera */
  renameCamera?: RenameCameraMessageProto | undefined;
  /** Calibration data for a camera */
  calibrationData?: CalibrationDataProto | undefined;
  /** Message to delete calibration data */
  deleteCalibration?: RemoveCalibrationDataMessageProto | undefined;
  /** Message to set camera recording status */
  setCameraRecordingStatus?: SetCameraRecordingStatusMessageProto | undefined;
  /** Message to set connection information for the device */
  setConnectionInfo?: SetConnectionInfoProto | undefined;
  /** Result of a pipeline operation */
  pipelineResult?: PipelineResultProto | undefined;
  /** Alert message containing alert data */
  alert?: AlertProto | undefined;
  /** List of pipeline types available in the system */
  pipelineTypeInfo: PipelineTypeProto[];
}